pipeline/requirements.txt
beautifulsoup4==4.14.3
resend
certifi==2026.2.25
charset-normalizer==3.4.5
google-generativeai==0.8.6
idna==3.11
python-dotenv==1.2.2
pytest==9.0.2
requests==2.32.5
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
